Move the PackageType Form Type class to the conventional path
