+++ src/Controller/DefaultController.php
@@ -11,7 +11,7 @@ namespace App\Controller;
 
 use App\Entity\Package;
 use App\Service\SitepackageGenerator;
-use App\Type\PackageType;
+use App\Form\PackageType;
 use App\Utility\StringUtility;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
